# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UPcar - ride sharing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [passenger, driver, admin]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DRIVER PROFILES ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── AUTH ────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        totalRides:
          type: integer
        accountStatus:
          type: string
          enum: [pending, approved, denied]
        createdAt:
          type: string
          format: date-time
        driverProfile:
          $ref: "#/components/schemas/DriverProfile"
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        avatarUrl:
          type: ["string", "null"]

    # ── DRIVER PROFILE ──────────────────────────────────────────────────────
    DriverProfile:
      type: object
      required: [id, userId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, denied]
        cpf:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        cnhNumber:
          type: ["string", "null"]
        cnhCategory:
          type: ["string", "null"]
        cnhExpiry:
          type: ["string", "null"]
        vehicleMake:
          type: ["string", "null"]
        vehicleModel:
          type: ["string", "null"]
        vehicleYear:
          type: ["integer", "null"]
        vehicleColor:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
          enum: [sedan, suv, hatch, pickup, moto, van, null]
        photoUrl:
          type: ["string", "null"]
        cnhPhotoUrl:
          type: ["string", "null"]
        vehiclePhotoUrl:
          type: ["string", "null"]
        criminalRecordUrl:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
...