# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * UPcar - ride sharing API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  phone?: string;
  /** @nullable */
  avatarUrl?: string | null;
}